+++ examples/UserStore.ts
@@ -1,21 +1,15 @@
-import { Store } from '../src/index'
-import { reactive } from 'vue'
+import { defineStore } from 'pinia'
 
-export default class UserStore extends Store {
-  protected prefix = 1
-  protected liveSync = true
-
-  state = reactive({
-    id: 0,
-    user: {
-      id: 0,
-      name: '',
-      age: 0
-    }
-  })
-
-  constructor() {
-    super()
-    super.init()
+export const useUserStore = defineStore('user', {
+  state: () => ({
+    favouritePosts: [1, 2],
+    token: 'asjdlkfjaZH:Hhkjhlkhk',
+  }),
+  actions: {
+  },
+  persist: {
+    sync: true,
+    suffix: 1,
+    name: 'user'
   }
-}
+})