+++ src/api/IApiResponse.ts
@@ -3,3 +3,9 @@ export interface IApiResponse<T> {
     count?: number,
     message?: string | string[]
 }
+
+export interface ApiResponse<T> {
+    data: T,
+    count?: number,
+    message?: string | string[]
+}