+++ src/collection/ICollection.ts
@@ -1,6 +1,6 @@
 import { IModelState } from '../model/IModelState'
-import { IQueryPage } from '../collection/IQueryPage'
-import { IQuery } from '../collection/IQuery'
+import { IQueryPage, QueryPage } from '../collection/IQueryPage'
+import { IQuery, Query } from '../collection/IQuery'
 import { IApi } from '../api/IApi'
 
 export interface ICollection<T> {
@@ -39,4 +39,44 @@ export interface ICollection<T> {
     queryString(): IQuery
 }
 
+export declare class Collection<T> {
+  data: T[]
+  api: IApi
+  state: IModelState
+
+  filter: any
+  include: string[]
+  attributes: string[]
+  fieldsSelection: string[]
+  paging: QueryPage
+  sorting: string[]
+
+  channel?: string
+
+  factory<T>(collection: T[]): void
+  get<T>(filter?: any): Promise<T[]>
+
+  where(filter: any): this
+  with(relationships: string[]): this
+  select(fields: string[]): this
+  sort(sorting: string[]): this
+  page(paging: IQueryPage): this
+
+  joinChannel(channel?: string): void
+  leaveChannel(): void
+  broadcastCreated(e: any): void
+  broadcastUpdated(e: any): void
+  broadcastDeleted(e: any): void
+
+  setStateLoading(): void
+  setStateSuccess(): void
+  setStateError(): void
+
+  updateDataSource<T>(data: T[]): void
+  queryString(): Query
+
+  fetching(payload?: any): void
+  fetchingError(err?: any): void
+  fetched(payload: any): void
+}