+++ src/collection/IQueryPage.ts
@@ -9,3 +9,15 @@ export interface IQueryPage {
     */
     limit?: number
 }
+
+export interface QueryPage {
+    /*
+    * The page number that the client is requesting
+    */
+    number?: number,
+
+    /*
+    * A limit on the number of resources to be returned
+    */
+    limit?: number
+}