+++ src/index.d.ts
@@ -1,15 +1,30 @@
-import { IApi } from './api/IApi'
-import { IAxiosError } from './api/IAxiosError'
-import { ICollection } from './collection/ICollection'
-import { IQuery } from './collection/IQuery'
-import { IQueryPage } from './collection/IQueryPage'
-import { IModel } from './model/IModel'
-import { IModelState } from './model/IModelState'
-import { IModelParams } from './model/IModelParams'
-import { IApiResponse } from './api/IApiResponse'
+import { Api, IApi } from './api/IApi'
+import { AxiosError, IAxiosError } from './api/IAxiosError'
+import { Collection, ICollection } from './collection/ICollection'
+import { IQuery, Query } from './collection/IQuery'
+import { IQueryPage, QueryPage } from './collection/IQueryPage'
+import { IModel, Model } from './model/IModel'
+import { IModelState, ModelState } from './model/IModelState'
+import { IModelParams, ModelParams } from './model/IModelParams'
+import { ApiResponse, IApiResponse } from './api/IApiResponse'
 import { BroadcastMessage } from './events/EventTypes'
+import { Permissions, Policy } from './policy/IPolicy'
 
 export {
+  Api,
+  ApiResponse,
+  AxiosError,
+  Collection,
+  Model,
+  ModelParams,
+  ModelState,
+  Permissions,
+  Policy,
+  Query,
+  QueryPage,
+  BroadcastMessage,
+
+  // Deprecated
   IApi,
   IApiResponse,
   IAxiosError,
@@ -18,6 +33,5 @@ export {
   IQueryPage,
   IModel,
   IModelParams,
-  IModelState,
-  BroadcastMessage
+  IModelState
 }