+++ src/model/IModelState.ts
@@ -3,3 +3,9 @@ export interface IModelState {
     isSuccess: boolean
     isError: boolean
 }
+
+export interface ModelState {
+    isLoading: boolean
+    isSuccess: boolean
+    isError: boolean
+}