+++ test/mocks/PostQueryApi.ts
@@ -1,9 +1,0 @@
-import { Api } from '../../src/index'
-
-export default class PostQueryApi extends Api {
-  protected resource = 'posts-query'
-
-  constructor() {
-    super()
-  }
-}